add, NIKL json parsing code

+++ Utils/dataloder.py
@@ -9,10 +9,10 @@ class ExoBrain_Datasets(Dataset):
         self.labels = np.load(path+"/labels.npy")
         self.token_type_ids = np.load(path+"/token_type_ids.npy")
 
-        self.attention_mask = torch.tensor(self.attention_mask)
-        self.input_ids = torch.tensor(self.input_ids)
-        self.labels = torch.tensor(self.labels)
-        self.token_type_ids = torch.tensor(self.token_type_ids)
+        self.attention_mask = torch.tensor(self.attention_mask, dtype=torch.int64)
+        self.input_ids = torch.tensor(self.input_ids, dtype=torch.int64)
+        self.labels = torch.tensor(self.labels, dtype=torch.int64)
+        self.token_type_ids = torch.tensor(self.token_type_ids, dtype=torch.int64)
 
     def __len__(self):
         return len(self.attention_mask)
